+++ proto/control-api/CHANGELOG.md
@@ -19,11 +19,13 @@ All user visible changes to this project will be documented in this file. This p
     - `ControlApi` and `CallbackApi` implementations ([#56]).
 - Cargo features:
     - `client` and `server` ([#56]);
-    - `serde` ([#67]).
+    - `serde` ([#67]);
+    - `client-api-proto` ([#68]).
 
 [#55]: /../../pull/55
 [#56]: /../../pull/56
 [#67]: /../../pull/67
+[#68]: /../../pull/68